Separate DynamicAvatar into its own file
